Catches missing instance
This commit aims to fix the case where we either don't get a response
back from ec2 or we have a timing issue that is causing nil to be
returned.  If we cannot find the instance by ipaddress we throw and
exception and skip the removal steps.  This is safe because we will
reload the list of instances and try again the next time through the
loop.

This fixes #5